requirements.txt
openai>=1.30.0
python-dotenv
sentence-transformers
scikit-learn
tqdm
pydantic
newsapi-python
requests
streamlit
altair
pandas
gtts
google-generativeai
